devices/media-center/keyboard2mqtt/requirements.txt
evdev==1.6.1
paho-mqtt==1.6.1
